requirements.txt
pytest~=8.0.0
allure-pytest
selene~=2.0.0rc7